requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==0.5.0
Django==4.0.6
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.3
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.11
whitenoise==6.2.0
django-filter~=22.1
pillow==9.2.0
django-simple-history==3.0.0
numpy~=1.23.2
tensorflow==2.9.1
opencv-python==4.6.0.66
keras~=2.9.0